sends completed orders to client

+++ order-processing-service/src/main/java/bove/order/processing/service/dto/order/OrderRequest.java
@@ -17,8 +17,8 @@ public class OrderRequest {
     private Double price;
     private String side;
     private String type;
-    private int portfolioId;
-    private int userId;
+    private Long portfolioId;
+    private Long userId;
     private String osId;
     private Date orderDate;
 }
